# perf(rome_js_formatter): Reduce the `String` allocations for Tokens