--- a/crates/rome_css_syntax/src/lib.rs
+++ b/crates/rome_css_syntax/src/lib.rs
@@ -4,7 +4,7 @@ mod syntax_node;
 
 pub use self::generated::*;
 pub use rome_rowan::{
-    SyntaxText, TextLen, TextRange, TextSize, TokenAtOffset, TriviaPieceKind, WalkEvent,
+    SyntaxNodeText, TextLen, TextRange, TextSize, TokenAtOffset, TriviaPieceKind, WalkEvent,
 };
 pub use syntax_node::*;